refactor: Break down subscription.go into modular files for better maintainability
- Created 7 focused modules from the monolithic 1933-line subscription.go:
  * manager.go: Core subscription manager types and global management
  * utils.go: Utility functions and relay mode detection
  * events.go: Kind 888 event creation and management
  * payment.go: Payment processing and tier matching logic
  * credit.go: Credit system management and automatic application
  * batch.go: Batch operations and scheduling for event updates
  * free_tier.go: Free tier management and daily renewal automation

- Preserved all existing functionality while improving code navigation
- Maintained subscription package structure and public interfaces
- Enhanced code organization for easier maintenance and feature development
