+++ lib/subscription/utils.go
@@ -1,0 +1,164 @@
+// utils.go - Utility functions and helpers
+
+package subscription
+
+import (
+	"fmt"
+	"log"
+	"strconv"
+	"strings"
+
+	"github.com/HORNET-Storage/go-hornet-storage-lib/lib/signing"
+	"github.com/nbd-wtf/go-nostr"
+	"github.com/nbd-wtf/go-nostr/nip19"
+	"github.com/spf13/viper"
+
+	"github.com/HORNET-Storage/hornet-storage/lib/types"
+)
+
+// normalizePubkey converts both npub and hex formats to a consistent format for comparison
+// Uses the signing package for decoding and go-nostr nip19 for npub encoding
+func normalizePubkey(pubkey string) (hex string, npub string, err error) {
+	// Use the signing package's DecodeKey which handles both hex and bech32
+	keyBytes, err := signing.DecodeKey(pubkey)
+	if err != nil {
+		return "", "", fmt.Errorf("invalid pubkey format: %v", err)
+	}
+
+	// Convert to hex format
+	hexKey := fmt.Sprintf("%x", keyBytes)
+
+	// Convert to npub format using go-nostr nip19 (consistent with rest of codebase)
+	npubKey, err := nip19.EncodePublicKey(hexKey)
+	if err != nil {
+		return "", "", fmt.Errorf("failed to encode as npub: %v", err)
+	}
+
+	return hexKey, npubKey, nil
+}
+
+// getRelayMode reads the current relay mode from configuration
+func (m *SubscriptionManager) getRelayMode() string {
+	// Load allowed users settings from config to get current mode
+	var allowedUsersSettings types.AllowedUsersSettings
+	if err := viper.UnmarshalKey("allowed_users", &allowedUsersSettings); err != nil {
+		log.Printf("[DEBUG] Warning: could not load allowed_users settings: %v", err)
+		return "unknown"
+	}
+
+	mode := strings.ToLower(allowedUsersSettings.Mode)
+	log.Printf("[DEBUG] Creating storage info for npub: mode=%s", mode)
+
+	// Validate mode and return
+	switch mode {
+	case "free", "paid", "exclusive", "personal":
+		return mode
+	default:
+		log.Printf("[DEBUG] Unknown mode '%s', defaulting to 'unknown'", mode)
+		return "unknown"
+	}
+}
+
+// getTagValue extracts a tag value by key from nostr event tags
+func getTagValue(tags []nostr.Tag, key string) string {
+	for _, tag := range tags {
+		if len(tag) > 1 && tag[0] == key {
+			return tag[1]
+		}
+	}
+	return ""
+}
+
+// getTagUnixValue extracts a unix timestamp from a tag (typically from index 2)
+func getTagUnixValue(tags []nostr.Tag, key string) int64 {
+	for _, tag := range tags {
+		if len(tag) > 2 && tag[0] == key {
+			unixTime, err := strconv.ParseInt(tag[2], 10, 64)
+			if err == nil {
+				return unixTime
+			}
+		}
+	}
+	return 0
+}
+
+// checkAddressPoolStatus checks if we need to generate more addresses
+func (m *SubscriptionManager) checkAddressPoolStatus() error {
+	availableCount, err := m.store.GetStatsStore().CountAvailableAddresses()
+	if err != nil {
+		return fmt.Errorf("failed to count available addresses: %v", err)
+	}
+
+	log.Println("Available count: ", availableCount)
+
+	// If we have less than 50% of addresses available, request more
+	if availableCount < 50 {
+		log.Printf("Address pool running low (%d available). Requesting 100 new addresses", availableCount)
+		return m.requestNewAddresses(20)
+	}
+
+	return nil
+}
+
+// getSubscriptionStatus returns the subscription status string
+func (m *SubscriptionManager) getSubscriptionStatus(activeTier string) string {
+	if activeTier == "" {
+		return "inactive"
+	}
+	return "active"
+}
+
+// findAppropriateTierForUser determines the appropriate tier for a user based on mode and current status
+func (m *SubscriptionManager) findAppropriateTierForUser(pubkey string, currentTier *types.SubscriptionTier, allowedUsersSettings *types.AllowedUsersSettings) *types.SubscriptionTier {
+	switch allowedUsersSettings.Mode {
+	case "free":
+		// In free mode, assign users to the first available free tier
+		for i := range allowedUsersSettings.Tiers {
+			if allowedUsersSettings.Tiers[i].PriceSats <= 0 {
+				return &allowedUsersSettings.Tiers[i]
+			}
+		}
+		// Fallback to basic free allocation
+		return &types.SubscriptionTier{
+			Name:              "Basic Free",
+			MonthlyLimitBytes: 100 * 1024 * 1024, // 100 MB
+			PriceSats:         0,
+			Unlimited:         false,
+		}
+
+	case "paid":
+		// In paid mode, give users the basic free tier by default
+		// They can upgrade by making payments
+		for i := range allowedUsersSettings.Tiers {
+			if allowedUsersSettings.Tiers[i].PriceSats <= 0 {
+				return &allowedUsersSettings.Tiers[i]
+			}
+		}
+		// No free tier configured, use minimal allocation
+		return &types.SubscriptionTier{
+			Name:              "Basic Free",
+			MonthlyLimitBytes: 100 * 1024 * 1024, // 100 MB
+			PriceSats:         0,
+			Unlimited:         false,
+		}
+
+	case "exclusive":
+		// In exclusive mode, tier is manually assigned
+		// Check if user is still in allowed lists
+		if m.isUserInAllowedLists(pubkey) {
+			if currentTier != nil {
+				return currentTier
+			}
+			// User is allowed but no tier assigned, give first available tier
+			if len(allowedUsersSettings.Tiers) > 0 {
+				return &allowedUsersSettings.Tiers[0]
+			}
+		}
+		// User not in allowed lists, no tier
+		return nil
+
+	default:
+		log.Printf("Unknown mode: %s", allowedUsersSettings.Mode)
+		return currentTier
+	}
+}
